// apps/billing/api/src/modules/reports/supplier-summary/supplier-summary.types.ts
export type SupplierSummaryItem = {
  balance: number;
  code: string;
  credit: number;
  debit: number;
  id: number;
  name: string;
};

export type SupplierSummaryResult = {
  companyId: number;
  companyName: string;
  financialYearId: number;
  financialYearName: string;
  items: SupplierSummaryItem[];
  total: number;
  totalBalance: number;
  totalCredit: number;
  totalDebit: number;
};